Whole-sale pasting the guts of Main into AppController
This leaves Program.Main really slim, handing off control ASAP to a separate class.
